Clean up access token refresh logic.
Refresh of token was a side effect of accessing the token property, which isn't
ideal.  This logic has been moved to a refresh_access_token() method, which can
be called on demand any time the token needs to be refreshed, or will be called
automatically when the token is expired.